+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix `al-table` throwing an error when receiving null data
 * Add `al-virtual-scroll` element
+* Make field `constraints` work by using the `Document` instance instead of plain data object
 
 ## 0.2.9 (2023-11-27)